+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.3] - 2022-06-30
+
 ### Added
 
 - Added status value 'Rejected' for Billing Statement entity
@@ -400,7 +402,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Entity Price: use PriceBand now, which is in the Catalog namespace where it belongs
 - Entity Service: the root entity is deprecated, and the one in the Catalog namespace should be used
 
-[Unreleased]: https://github.com/ArrowSphere/public-api-client/compare/0.9.2...HEAD
+[Unreleased]: https://github.com/ArrowSphere/public-api-client/compare/0.9.3...HEAD
+[0.9.3]: https://github.com/ArrowSphere/public-api-client/compare/0.9.2...0.9.3
 [0.9.2]: https://github.com/ArrowSphere/public-api-client/compare/0.9.1...0.9.2
 [0.9.1]: https://github.com/ArrowSphere/public-api-client/compare/0.9.0...0.9.1
 [0.9.0]: https://github.com/ArrowSphere/public-api-client/compare/0.8.0...0.9.0